[Core] Fix class name (If --> Id)
